[MOV] mrp: move `stock.move.line` is its own file
> Each inherited model should be in its own file to help understanding
of impacted models.

https://www.odoo.com/documentation/master/contributing/development/coding_guidelines.html#file-naming
